Added POST tests
This work also consolidated the test code for better sharing.

Also switched to using httpbin.org/status/ for creating the responses
which seems to be a more reliable source. The status messages that come
back from this server are upper-case, so made a change to make the test
msg comparison case-insensitive (which is what a lot of the WWW is like)